+++ simpleDS/tests/test_delay_spectrum.py
@@ -10,16 +10,19 @@ import numpy as np
 import copy
 import nose.tools as nt
 import unittest
-from simpleDS import DelaySpectrum, delay_spectrum as dspec
-from simpleDS import utils
-from simpleDS.data import DATA_PATH
-from pyuvdata.data import DATA_PATH as UVDATA_PATH
+
 from pyuvdata import UVBeam, UVData
 import pyuvdata.tests as uvtest
 from astropy import constants as const
 from astropy import units
 from astropy.cosmology import Planck15
 from scipy.signal import windows
+
+from simpleDS import DelaySpectrum, delay_spectrum as dspec
+from simpleDS import utils
+from simpleDS.data import DATA_PATH
+from pyuvdata.data import DATA_PATH as UVDATA_PATH
+import simpleDS.tests as sdstest
 
 
 class TestClass(object):
@@ -1073,3 +1076,25 @@ def test_update_cosmology_unit_and_shape_uncalib():
 
     dspec_object.update_cosmology(cosmology=test_cosmo)
     nt.assert_true(dspec_object.check())
+
+
+@sdstest.skipIf_py2
+def test_update_cosmology_littleh_units():
+    """Test the units can convert to 'littleh' units in python 3."""
+    testfile = os.path.join(UVDATA_PATH, 'test_redundant_array.uvh5')
+    test_uvb_file = os.path.join(DATA_PATH, 'test_redundant_array.beamfits')
+    test_cosmo = Planck15
+    uvd = UVData()
+    uvd.read(testfile)
+    dspec_object = DelaySpectrum(uv=[uvd])
+    dspec_object.select_spectral_windows([(1, 3), (4, 6)])
+    uvb = UVBeam()
+    uvb.read_beamfits(test_uvb_file)
+    dspec_object.add_uvbeam(uvb=uvb)
+    dspec_object.calculate_delay_spectrum()
+    nt.assert_true(dspec_object.check())
+
+    dspec_object.update_cosmology(cosmology=test_cosmo, littleh_units=True)
+    nt.assert_true(dspec_object.check())
+    test_unit = (units.mK**2) / (units.littleh / units.Mpc)**3
+    nt.assert_equal(dspec_object.power_array.unit, test_unit)